Bot: Record gas usage for all addresses.
We need to set threshold on the gas usage, in case it goes unexpectedly quickly. To set the threshold, we need some baseline. This change records the last 1000 records of gas left in wallets.

The data are stored in `App::gases` as `HashMap<Address, Vec<u128>>`, updated in every `gas_check` run. The data are shown in `/debug/gases` endpoint, just `Debug` output, no format, no prettify.