Update Twitter card metadata
I was seeing some instability in getting Twitter to show the image in
the card. Ultimately, I don't think it had anything to do with us, but
I'm not certain. In any event, it led me down a rabbit hole of
optimizing the Twitter share data so it can be used on other sites as
well.

Typed URLs made this really nice :)