Integrate Sentry to bots
In current Ops design, services send alerts to Sentry then triggers other function parts. Also Sentry could do some analyzation on error history.

To avoid the quota issue, reduce the messages sent to Sentry, this implenmentation follows `p2s`, only sends state change messages to Sentry, which means on error first occurs, and on error is gone.

The definition of "error occur" is the final `Err` `Result` of a task, after it retries.

The definition of "error gone" is that last time, the task failed, and this time, the task succeeded.

When a bot restarted, and a task runs for the first time and succeeds, an info message will be sent indicating all previous errors should be considered recovered.