Add pid1 image to the bots dockerfile
When we deploy a new task definition using Amazon ECS, it is taking
quite a bit of time for the new task to come up to life.

When ECS stops a running task, it sends a SIGTERM signal. Right now it
takes quites a bit of time which indicates that the signal is likely
being ignored. Using pid1, will hopefully solve this issue.