perp-1985 | migrate the Pyth bridge to Hermes
This operates by checking if we're on the old version of the pyth_bridge
and, if so, instantiating a fresh bridge and then updating the price
admin. Tested on osmodebug and it worked just fine.

Once this is merged in, and once the frontend changes for Hermes are
deployed to production, we can begin migrating all the other testnet
contracts and updating their bots. Then the final step will be doing
this migration process for mainnet. In that case, I'll probably do the
migration steps more manually.