Do tracing for crank and price updates
While this does work, I have found the sentry client to be *not* very
reliablie.

There are two ways to do distributed traces via sentry:
- Automatic tracing via macro (This is how we have done it here)
- Tracing by writing code. This is very complex and error prone to do
since an await can yield and switch to another task. And it wil
include span for the other task. This would make refactoring also
quite complicated as all the await point needs to be audited for
correct code.

Given the situation, I have gone with automatic tracing. Now the cons
of automatic tracing: It doesn't work reliably. I have observed that
it stops dropping data after some time making it less helpful. My
guess is that this is likely a bug in the sentry-tracing crate.