Drop OnceCell for StatusMap
This is significantly more involved than #331, which also fixes the bug,
so it's debatable whether this is worth merging (please feel free to
push back). The idea though is to be much more explicit about
transferring ownership of the StatusMap to the REST server so that we
don't even need a mutable cell.