perp-2386 | add a suppress for task output
This moves WatchedTaskOutput to use a builder pattern, which creates a
lot of noise in the PR. The core concept is adding a new suppress field
which, when true, will avoid overwriting old statuses. The purpose for
now is that, when doing crank run, we will not wipe out the useful
information of the last transaction sent with useless "nothing
interesting happened" messages.