perp-2754 | ignore errors after epoch
This is a slightly hacky approach, since (1) it relies on the bots having been running during the epoch to populate a variable and (2) it also relies on an error having occured during that period. I think this will end up being Good Enough in all cases we care about, and is much easier to implement than other approaches I could think of, so I wanted to try this out first and confirm it works as expected.

This should _not_ get deployed before the weekend, it should only go live Sunday the earliest and we should manually confirm that it works as expected at the next epoch.