During liquidation trigger, liquifund till next liquifunding
Opening as a separate PR instead of continuing discussion on the previous one. I believe this is the correct way to handle things, but I'm not certain. We want to first do a liquifunding for the entire liquifunding period. We need to use next_liquifunding to do that correctly, since we may have a fuzz factor in place.

The other change is close_time. Previously it was using the ends_at, which was the calculated liquifunding point. However, I believe we should instead be using the current block time, which as I understand it is the point of the field.

Hopefully we'll get better clarity on "close_time" vs "settlement_time" when working on PERP-2818.