perp-2953 | initial milkTIA config
milkTIA uses the same contracts as Stride, so we can simply reuse that
code. It did introduce a new concept of market-specific Stride contract
addresses. For now, I've addressed this by having a "fallback" and
"override" value, but it's also an option to simply remove the fallback
and be forced to provide a Stride address per-market.