Tweaks to trading incentives program
These came up while validating the data against manual calculation in Excel. There was one issue I hadn't picked up on that definitely led to incorrect results: we need to calculate only trading fees, not total fees. Total fees includes funding payments, which will screw results up.

Outside of that, I ended up doing a significant rewrite of the algorithm. Personally I think it's cleaner, but reality is I'm not sure if the changes impacted the behavior or not, since I was having a lot of trouble reconciling with Excel. In any event, values are now reconciled. Summary of changes:

* Did a streaming processing of CSVs
* Introduced the TotalsTracker
* Renamed some variables and parameters, and provided some useful defaults
* Added a `type` column to the output to simplify some testing