Fix countertrade::closes_extra_positions test
The test used to fail because it tried to close all the position,
forced opened by the countertrade address.

Specifically this line used to fail:

// Nothing left to be done now
assert_eq!(
    market.query_countertrade_has_work().unwrap(),
    HasWorkResp::NoWork {}
);

Instead the response that came was instructing the contract to close
position 5.

And looking at the contract code, that's because this condition was
true:

popular_funding < min_funding

Ref:
https://github.com/Levana-Protocol/levana-perps/blob/6f7ab34bc021c666e7776635adf315b5ac92bef5/contracts/countertrade/src/work.rs#L180

In the test case, these were it's values:

popular_funding = 0.049
min_funding = 0.1

For fixing the tests, I have changed tweaked the short position's
collateral and leverage to be such that the market is balanced and no
new work is required.