Redo of CopyTradingInfo stuff
Probably overall makes sense to standardize on using CopyTradingInfo instead of a tuple in the types, that would also allow for the reference-type optimization that bypasses some clones. But sticking to a pair of Addrs is the simpler approach.

Note, I ran out of time before my next call to get everything lined up, but took care of a few of the error cases.