perp-4165 | handle the missing final closed position bug
This is especially noticeable when using a limit of 1, but could in theory happen elsewhere also. I changed the logic to use exclusive bounds and to make the cursor represent the _next_ position we'll see, which is more in line with how other parts of the code do enumerations. There are other ways to skin this, but this seems like the most direct. Downside: slight change in what users will receive. We should confirm through manual testing that all positions still appear in the frontend with this change.