Remove deferred execution queueing check
This is because we already track all the deferred exec ids that is
produced in the copy trading contract. We don't proceed with new work
items unless the pending deferred exec ids are cranked.

So there is no need for us to query it in the get_work step. This
avoid a huge amount of smart query calls.